Add router loading from JSON and refactor JsonRouterLoader
A method for loading routes from a JSON file was added to the Router class. The JsonRouterLoader was refactored to simplify its use with notable changes including making it a static class and adding a method to check the format of the routes from a JSON file. A test controller was also added and the routes.json file was updated.