Add Nix installation script and update Makefile
The commit introduces a new installation script for the Nix package manager, which checks whether curl and Nix are installed before proceeding with the Nix installation. The Makefile has also been updated to call this script for the `install-nix` target. A new `run-prod` target has been added to the Makefile as well.