Update ReactHttpServer class and add environment configurations
This update refactors the ReactHttpServer class to make it more flexible and dynamic. The changes include flexible server initialization with host and port parameters and dynamic route loading from a specified JSON file. Additionally, vlucas/phpdotenv was added to the project for loading environment variables to the server. Several unused uses statements were removed for cleaning purposes. The .env file was also introduced with the path for the routes JSON file specified. Overall, this improves the configurability of the server and code cleanliness.