Refactor HtmlController and update routes.json
HtmlController's invoker method was refactored to use a construct method and perform additional checks for file existence and type. The HTML resource's $uri was also updated to return an empty string by default. In addition, the reference paths for HealthController and FindPostByIdController in routes.json have been updated to reflect their correct paths.