Update HTML pages, README.mb, scripts and introduce Phinx
This commit updates the design of HTML pages by adding a background and completes relocation of script files. It also introduces the database migration tool Phinx into the project and adds detailed usage instructions to README.mb. The README.mb description section is rewritten for better clarity and a brand new index HTML page is created for sleeker presentation.