Update ConnectionPool class methods and add functionalities
Updated 'attach' and 'detach' methods to 'add' and 'remove' in the ConnectionPool class for better semantics. Added the ability to assign UUIDs to each connection, as well as methods to find connections by UUID, exclude a specific connection from messaging, and send data to all connections. Made corresponding modifications to the Chat class.