Add ExceptionTo class and refactor exception handling
An ExceptionTo class has been added to encapsulate the logic for converting an exception into an array format. This class has been used in JsonResponse.php for replacing previous inline array transformations. The substantial benefit is that exceptions are now refined into a JSON form that is easier to analyze in a browser response.