Refactor websocketTest.html for better code structure
This commit revises the websocketTest.html file by restructuring some of the JavaScript used for setting up the ChatBoxComponent. It introduces the $ shorthand for document.querySelector and embeds the chat setup within a function to improve clarity and organization. The IP address for the chat host attribute has also been updated.