Update color definitions in CSS and theme switcher
The CSS styling of `pascualmg-portfolio.js` and color definitions in `themeSwitcher.js` have been updated. Changes include adjusting the header's background color and adding color styles for h2 and h3 in `pascualmg-portfolio.js`, and changing the `head1` color definitions in `themeSwitcher.js`. These alterations lead to improved readability and customization of the UI.