Replace post id with UUID in Post entity
This change modifies the post entity to use a UUID instead of an integer id for better uniqueness and scalability. The post table migration was also updated to use a string of 36 characters for the id column instead of an integer. A new `Uuid` value object was implemented using the Ramsey Uuid library, which is then used in the post entity to represent the id.