Refactor HTML and CSS, implement PostCreator domain service
This commit mainly focuses on refactoring the HTML and CSS of how-to-create-a-new-post. It includes replacing the manual style definitions with a link to the spacemacs.css stylesheet. Furthermore, an instance of the PostCreator domain service is implemented, which is in charge of creating new posts utilizing the PostRepository. This will improve code clarity and reusability while maintaining the same visual appearance of the site.