Add project structure to README
A new section has been added to the README which details the structure of the project. This lays out the project according to Domain-Driven Design (DDD) principles. The section also explains the role of each directory within the src/ddd segment of the project. Furthermore, the utility of the Github template for the project is also discussed.