Enhance project structure detail in README file
The project's README.org has been updated to provide a more precise structure of the Cohete project. It details the new file hierarchy within the project, focusing mainly on the 'src/ddd' directory and its importance in Domain-Driven Design (DDD). The file paths, functionality, and relations to the project as a whole are explained. The crucial roles of 'Kernel.php' and 'ReactHttpServer.php' are also highlighted, guiding users to understand Cohete's core application function, and its reactive HTTP server.