Refactor project hierarchy documentation in README
The structure documentation for the Cohete project in README.org has been updated, by focusing on a more precise file hierarchy explanation. The update particularly spotlights on the 'ddd' directory and its fundamental role in Domain-Driven Design (DDD). The revision underscores the crucial 'Kernel.php' and 'ReactHttpServer.php' files, providing a comprehensive guide towards understanding the core of the application and the reactive HTTP server.